Add compatibility with Django1.7.
In django 1.7, we might have some AppConfigs in the INSTALLED_APPS, and
not only modules. This was breaking our (ugly) way of autodiscovering
"mails.py" modules.

We now use the much cleverer and cleaner "app-loading" way available in
Django1.7.

Also, to take full advantage of the autodiscover_modules utility, the
`MailFactory.mail_map` attribute has been renamed to `_registry`.