Support configuring HTML2Text for plain text rendering
The plain text generated by `HTML2Text` doesn't look very good.
Especially when you are using table heavy HTML emails. So with this
addition one should be able to configure how the plain text is
generated by configuring it. Simply add the setting
`MAIL_FACTORY_HTML2TEXT` as a dictionary that maps to the options on
`Html2Text` class.